osint/requirements.txt
opencv-python
librosa
numpy
flask
requests
beautifulsoup4
snscrape
Pillow
exifread
networkx
matplotlib
reportlab
torch
torchvision
torchaudio
dnspython